+++ apps/test/unit/lib/kits/maker/MicroBitComponentsTest.js
@@ -4,7 +4,7 @@ import {
   createMicroBitComponents,
   cleanupMicroBitComponents,
   enableMicroBitComponents
-} from '@cdo/apps/lib/kits/maker/MicroBitComponents';
+} from '@cdo/apps/lib/kits/maker/boards/microBit/MicroBitComponents';
 import {MicrobitStubBoard} from './makeStubBoard';
 import five from '@code-dot-org/johnny-five';
 import sinon from 'sinon';
